IMSIPseudo.java: import all
It takes time to figure out where exactly a certain class/object/... is
located. Just import all of them. It does not seem to impact the size of
the applet, as there's a blog post writing about performance that does
it the same way [1]. Besides, except for the HelloSTK hello world, I've
seen it in all other code done that way, including 3GPP 03.19 Annex C.

[1]: https://gelvaos.blogspot.com/2011/04/