IMSIPseudo.java: fix up length somewhat
Since it is counting nibbles, we need to take the length, which is in
bytes, two times. Also increase the size of the whole string, because
there are additional numbers infront of the imsi, it seems.

There are still two numbers missing at the end, and there is a
mysterious ) character. But other than that, it's working.