spec: SIM applet: verify: next IMSI is different
Suggested by ROS, so it would not be possible for the HLR to send the
same pseudonymous IMSI with an increased counter.

Seems like a good idea to me, if there was a bug in the HLR and it would
send the same pseudo IMSI for that reason, the SIM would not reset the
imsi_pseudo_lu counter.